// src/app/enums/card11-key.enum.ts
export enum Card11KeyEnum {
  Lang = 'lang',
  Sell = 'sell',
  Coin = 'coin',
  Factor = 'factor',
  Pnl = 'pnl',
  Pnl2 = 'pnl2',
  Background = 'background',
  AvgEntryPrice = 'fair-price',
  AvgClosePrice = 'entry-price',
  OpenTime = 'time',
  CloseTime = 'close_time',
  Referral = 'referral',
  Qr = 'qr',
  AvgEntryPriceTitile = 'fair-pricet',
  AvgClosePriceTitile = 'entry-pricet',
  OpenTimeTitile = 'timet',
  CloseTimeTitile = 'close_timet',
}